backend/requirements.txt
fastapi==0.135.2
uvicorn==0.42.0
SQLAlchemy==2.0.48
psycopg2-binary==2.9.11
pydantic==2.12.5
email-validator==2.3.0
passlib==1.7.4
bcrypt==4.0.1
python-jose==3.5.0
python-dotenv==1.2.2
httpx==0.28.1
apscheduler
swagger-ui-bundle==1.1.0
